Citation: major contribution by Sascha Reißner.
Now all circuit-breakers exist and are ready to wire up.
The circuit-breakers for the engine instuments are wired.

The engine instruments are completely reworked.
Every tape has his own filter and behave like in real live.
Fan speed work without power, if the fan is over 50%.
Otherwise it falls to zero.
The other ones are falling under zero and show a OFF-flag.
Turbine digits are now flashing if at or over 96%.

Colors of the annunciation panel corrected.
All annunciators exist and are ready to wire up.

I found 'controls/gear/antiskid', but there is now script for that.
It is probably from the core. Anyway, i have activated the antiskid
schwitch and wired up (over left bus and breakers).
I dont know if it is working.
At gear extantion, the [ANTI-SKID INOP] will lit for 4sec.
(Antiskid runs atest everytime the gears are extented)
It will also lit, if the switch in OFF position.

Everytime a hydraulic device (gear, speedbrake, reverser) is used,
the [HYD PRESS ON] annunciator will lit.

The throttle lever have now release trigger.
This prevents for inadvertent open the cut-off valve in cold position
and for inadvertent cut-off while flying.
To release, pull up the trigger. If you move the lever forward to IDLE
or backward to OFF, the detent will spring loaded move down and lock
the lever.
Because the triggers are hard to pick, there is a 'End' key binding.

The throttle lever has also a trust reverser lever (Citation II only).
Trust reverser is only activatable in IDLE position and if the
maingear-switch is closed. The maingear-switch is a switch on the left
gear torqe link and close if the gear is under compression.
You can find it under 'controls/electric/maingear-switch'.
This prevents for inadvertent activation while flying.
If you activate the reverser, the reverser lever will pop up.
As long as the reverser is active, thrust will move the reverser lever
up.The throttle lever will remain on IDLE.

To shut down the engines, lift the release trigger and move the
throttle back to OFF.

To start the engines, push the starter button.
The start circuit will automaticly activate the boost pump.
Lift the release trigger.
If the turbine at 8-10%, move the throttle up to IDLE.
The start circuit will automaticly activate the ignition.
If turbine is on 20% and the fan isn't come up (more then 5%),
the start circuit will automaticly disengaged.
If the generator is ON while starting, the annunciator will show that
it is off. This is because the generator works as starter.
The generator will come on line after the engine is started.

And there are other little glitches that i cant remember anymore. ;)

I wish habby flying and a nice Xmas.

git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/Citation@3155 9eb9bc84-db9d-4e05-893b-26dd41ac94a6