Citation version 4.0.0: new multiplayer protocol and some bug fixes
The multiplayer protocol in this version is incompatible with previous versions of the aircraft
but this only affects people who have renamed /usr/share/games/flightgear/AI/Aircraft/Citation
in order to use the detailed model to see remote Citations.

* Set /sim/multiplay/protocol-version to 2: require at least fgfs 2017.2.
* Use absolute property names of the local fgfs for rendering, i.e. /sim/rendering/*, even in
  animations of a remote Citation.
* Use normal property names in all animations instead of the sim/multiplay/generic/ ones.
* Models/Citation-II.xml (nasal/load):
  - alias the sim/multiplay/generic/ properties to the normal ones.
  - Add the following properties to the protocol:
    controls/human-models: attempt to show humans over mp.
    gear/gear[0..2]/rollspeed-ms: animate wheels over mp.
    gear/gear[0..2]/compression-m: animate compression over mp.
  - Use short (4 bytes incl. tag) instead of int (8 bytes incl. tag) for voltage properties,
    systems/electrical/outputs/*.

The attempt to show humans over mp fails for now because the animation requires additional
properties not initialized in the property subtree of the remote Citation, namely
/sim/model/crew/pilot[0..1]/*, /sim/model/pax[0..6]/*.  They are declared in
Citation-II-common.xml, which is not loaded when the Citation is a remote aircraft (only
Models/Citation-II.xml is loaded).  This probably affects the walker, too.

git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/Citation@3256 9eb9bc84-db9d-4e05-893b-26dd41ac94a6